xxx

xxxx

Super class

Magellan::ScreenManager -> Pipeline

Public fields

tmp.return

xxx

modal_txt

xxx

Methods

Public methods

Pipeline$ToggleState_Screens()

Pipeline$Discover_Skipped_Steps()

Pipeline$Set_All_Reset()

Pipeline$ValidateCurrentPos()

Pipeline$Additional_Server_Funcs()

Pipeline$ActionOn_NewPosition()

Pipeline$EncapsulateScreens()

Pipeline$GetScreens_ui()

Pipeline$ActionOn_New_DataIn()

Pipeline$Launch_Module_Server()

Pipeline$ActionOn_Data_Trigger()

Pipeline$GetMaxValidated_BeforeCurrentPos()

Pipeline$GetMaxValidated_BeforePos()

Pipeline$PrepareData2Send()

Pipeline$clone()

Inherited methods

Magellan::ScreenManager$Change_Current_Pos()

Magellan::ScreenManager$Get_Result()

Magellan::ScreenManager$Set_Reseted()

Magellan::ScreenManager$Set_Skipped()

Magellan::ScreenManager$initialize()

Magellan::ScreenManager$server()

Magellan::ScreenManager$ui()

Method ToggleState_Screens()

xxx

xxx

Usage

Pipeline$ToggleState_Screens(cond, range)

Arguments

cond

xxx

range

xxx

Returns

Nothing

Method Discover_Skipped_Steps()

xxx

Usage

Pipeline$Discover_Skipped_Steps()

Returns

Nothing

Method Set_All_Reset()

xxx

Usage

Pipeline$Set_All_Reset()

Returns

Nothing

Method ValidateCurrentPos()

xxx

Usage

Pipeline$ValidateCurrentPos()

Returns

Nothing

Method Additional_Server_Funcs()

xxx

Usage

Pipeline$Additional_Server_Funcs()

Returns

Nothing

Method ActionOn_NewPosition()

xxx

Usage

Pipeline$ActionOn_NewPosition()

Returns

Nothing

Method EncapsulateScreens()

xxx

Usage

Pipeline$EncapsulateScreens()

Returns

Nothing

Method GetScreens_ui()

xxx

Usage

Pipeline$GetScreens_ui()

Returns

Nothing

Method ActionOn_New_DataIn()

xxx

Usage

Pipeline$ActionOn_New_DataIn()

Returns

Nothing

Method Launch_Module_Server()

This function calls the server part of each module composing the pipeline

Usage

Pipeline$Launch_Module_Server()

Returns

Nothing

Method ActionOn_Data_Trigger()

Catch the return value of a module and update the list of isDone modules This list is updated with the names of datasets present in the rv$tmp variable. One set to TRUE all the elements in isDone which have a corresponding element in names(rv$tmp). One cannot simply set to TRUE the last element of rv$tmp because it will does not work in case of a reseted module (it is not in the names(rv$tmp) list anymore) If a value (not NULL) is received, then it corresponds to the module pointed by the current position This function also updates the list isDone This function updates the current dataset (self$rv$dataIn)

Usage

Pipeline$ActionOn_Data_Trigger()

Returns

Nothing

Method GetMaxValidated_BeforeCurrentPos()

This function calls the server part of each module composing the pipeline

Usage

Pipeline$GetMaxValidated_BeforeCurrentPos()

Returns

Nothing

Method GetMaxValidated_BeforePos()

This function calls the server part of each module composing the pipeline

Usage

Pipeline$GetMaxValidated_BeforePos(pos)

Arguments

pos

xxx

Returns

Nothing

Method PrepareData2Send()

This function calls the server part of each module composing the pipeline

Usage

Pipeline$PrepareData2Send()

Returns

Nothing

Method clone()

The objects of this class are cloneable with this method.

Usage

Pipeline$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.