xxx

xxxx

Super class

Magellan::ScreenManager -> Process

Public fields

modal_txt

xxx

Methods

Public methods

Process$ToggleState_Screens()

Process$Set_All_Skipped()

Process$Discover_Skipped_Steps()

Process$Set_All_Reset()

Process$ValidateCurrentPos()

Process$EncapsulateScreens()

Process$GetScreens_ui()

Process$clone()

Inherited methods

Magellan::ScreenManager$ActionOn_NewPosition()

Magellan::ScreenManager$ActionOn_New_DataIn()

Magellan::ScreenManager$Additional_Server_Funcs()

Magellan::ScreenManager$Change_Current_Pos()

Magellan::ScreenManager$Get_Result()

Magellan::ScreenManager$Set_Reseted()

Magellan::ScreenManager$Set_Skipped()

Magellan::ScreenManager$initialize()

Magellan::ScreenManager$server()

Magellan::ScreenManager$ui()

Method ToggleState_Screens()

xxx

xxx

Usage

Process$ToggleState_Screens(cond, range)

Arguments

cond

A number

range

A number

input

A number

output

A number

Method Set_All_Skipped()

Set to skipped all steps of the current object

Usage

Process$Set_All_Skipped()

Returns

Nothing

Method Discover_Skipped_Steps()

xxx

Usage

Process$Discover_Skipped_Steps()

Returns

Nothing

Method Set_All_Reset()

xxx

Usage

Process$Set_All_Reset()

Returns

Nothing

Method ValidateCurrentPos()

xxx

Usage

Process$ValidateCurrentPos()

Returns

Nothing

Method EncapsulateScreens()

xxx

Usage

Process$EncapsulateScreens()

Returns

Nothing

Method GetScreens_ui()

xxx

Usage

Process$GetScreens_ui()

Returns

Nothing

Method clone()

The objects of this class are cloneable with this method.

Usage

Process$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.