xxx

xxxx

Public fields

id

xxx

ns

xxx

verbose

xxx

currentStepName

xxx

child.process

xxx

length

xxx

original.length

xxx

config

xxx

screens

xxx

modal_txt

xxx

timeline

xxx

default_pos

xxx

dataOut

xxx

rv

xxx

Methods

Public methods

ScreenManager$new()

ScreenManager$ActionOn_New_DataIn()

ScreenManager$Additional_Server_Funcs()

ScreenManager$Set_Skipped()

ScreenManager$Set_Reseted()

ScreenManager$ValidateCurrentPos()

ScreenManager$EncapsulateScreens()

ScreenManager$ActionOn_NewPosition()

ScreenManager$Get_Result()

ScreenManager$Change_Current_Pos()

ScreenManager$ToggleState_Screens()

ScreenManager$ui()

ScreenManager$server()

ScreenManager$clone()

Method new()

Change current position.

Default actions on reset pipeline or process.

Set widgets of all screens to their default values.

Gives the name of the status corresponding to the code (integer).

Returns the date and time in timestamp UNIX format.

xxxx

xxx

xxx

xxx

xxx

xxx

Return the UI for a modal dialog with data selection input. If 'failed' is TRUE, then display a message that the previous value was invalid.

xxx

xxx

xxx

xxx

xxx

Usage

ScreenManager$new(id, verbose)

Arguments

id

xxx

verbose

xxx

direction

xxx

conf

A list containing the configuration of the current object. See xxx

name

A number

cond

xxx

range

xxx

cond

xxx

input

xxx

output

xxx

input

xxx

output

xxx

Method ActionOn_New_DataIn()

xxx

Usage

ScreenManager$ActionOn_New_DataIn()

Method Additional_Server_Funcs()

xxx

Usage

ScreenManager$Additional_Server_Funcs()

Method Set_Skipped()

xxx

Usage

ScreenManager$Set_Skipped()

Method Set_Reseted()

xxx

Usage

ScreenManager$Set_Reseted()

Method ValidateCurrentPos()

xxx

Usage

ScreenManager$ValidateCurrentPos()

Method EncapsulateScreens()

xxx

Usage

ScreenManager$EncapsulateScreens()

Method ActionOn_NewPosition()

xxx

Usage

ScreenManager$ActionOn_NewPosition()

Method Get_Result()

xxx

Usage

ScreenManager$Get_Result()

Method Change_Current_Pos()

xxx

Usage

ScreenManager$Change_Current_Pos(i)

Arguments

i

xxx

Method ToggleState_Screens()

xxx

Usage

ScreenManager$ToggleState_Screens(cond, range)

Arguments

cond

xxx

range

xxx

Method ui()

xxx

Usage

ScreenManager$ui()

Method server()

xxx

Usage

ScreenManager$server(dataIn = reactive({
    NULL
}))

Arguments

dataIn

xxx

Method clone()

The objects of this class are cloneable with this method.

Usage

ScreenManager$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.