R code to update the 'config' variable of a process module

This function generates the necessary code to modify the variable 'config' (slots steps and mandatory). It adds a 'Description' step and a TRUE value at the beginning of the 'steps' and 'mandatory' list, erases all white spaces for the names of the steps.

This function generates the necessary code to modify the variable 'config' (slots steps and mandatory). It adds a 'Description' step and a TRUE value at the beginning of the 'steps' and 'mandatory' list, erases all white spaces for the names of the steps.

This function create the source code needed inside a module to declare the reactive variable called 'widgets.default.values'. # Declaration of the variables that will contain the values of the widgets To avoid confusion, the first string is the name of the step while the second is the name of the widget

This function create the source code needed inside a module to declare the reactive variable called 'rv.custom.default.values'. # Declaration of the variables that will contain the values of the user variables. To avoid confusion, the first string is the name of the step while the second is the name of the widget

This function xxx # Generate dynamically the observeEvent function for each widget

This function createxxx

Returned value of the process

xxx

This function xxx # Generate dynamically the observeEvent function for each widget

This function xxx # Generate dynamically the observeEvent function for each widget

This function xxx # Generate dynamically the observeEvent function for each widget

This function xxx # Generate dynamically the observeEvent function for each widget

Get_Code_Update_Config_Variable()

Get_Code_Declare_widgets(widgets.names = NULL)

Get_Code_Declare_rv_custom(rv.custom.names = NULL)

Get_Code_for_ObserveEvent_widgets(widgets.names = NULL)

Get_Code_for_rv_reactiveValues()

Get_Code_for_dataOut()

Get_Code_for_General_observeEvents()

Module_Return_Func()

Get_Worflow_Core_Code(w.names = NULL, rv.custom.names = NULL)

Get_AdditionalModule_Core_Code(w.names = NULL, rv.custom.names = NULL)

Get_Code_for_AddMod_observeEvents()

Arguments

widgets.names

A `list` containing the names of the widgets in all steps of the module.

rv.custom.names

xxx

w.names

xxx

Value

A `string` containing some R code

NA

NA

NA

A `string` containing some R code

NA

NA

NA

NA

NA

NA

Author

Samuel Wieczorek

Examples

Get_Code_Update_Config_Variable()
#> [1] "\n    config@steps <- setNames(config@steps,\n        nm = gsub(' ', '', config@steps, fixed = TRUE))\n\n\n    "

widgets <- paste0("widget", 1:3)
Get_Code_Declare_widgets(widgets)
#> [1] "rv.widgets <- reactiveValues(\n\twidget1 = widgets.default.values$widget1,\n\twidget2 = widgets.default.values$widget2,\n\twidget3 = widgets.default.values$widget3\n)\n\n"

custom <- list(foo1 = list(), foo2 = 3)
Get_Code_Declare_rv_custom(names(custom))
#> [1] "rv.custom <- reactiveValues(\n\tfoo1 = rv.custom.default.values$foo1,\n\tfoo2 = rv.custom.default.values$foo2\n)\n\n"

widgets <- paste0("widget", 1:3)
Get_Code_for_ObserveEvent_widgets(widgets)
#> [1] "observeEvent(input$widget1, {\n    rv.widgets$widget1 <- input$widget1})\nobserveEvent(input$widget2, {\n    rv.widgets$widget2 <- input$widget2})\nobserveEvent(input$widget3, {\n    rv.widgets$widget3 <- input$widget3})\n\n\n"

Get_Code_for_rv_reactiveValues()
#> [1] "rv <- reactiveValues(\n    # Stores the object given in input of the process\n    dataIn = NULL,\n    # A vector of boolean indicating the status (UNDONE, SKIPPED or VALIDATED)\n    # of the steps\n    steps.status = NULL,\n    # xxx\n    reset = NULL,\n    # A vector of boolean indicating if the steps are enabled or disabled\n    steps.enabled = NULL\n  )\n\n    "

Get_Code_for_dataOut()
#> [1] "dataOut <- reactiveValues(\n                trigger = NULL,\n                value = NULL\n                )\n\n"

Get_Code_for_General_observeEvents()
#> [1] "observeEvent(steps.enabled(), ignoreNULL = TRUE, {\n    if (is.null(steps.enabled()))\n        rv$steps.enabled <- setNames(rep(FALSE, rv$length),\n                                 nm = names(rv$config@steps))\n    else\n        rv$steps.enabled <- steps.enabled()\n})\n\nobserveEvent(steps.status(), ignoreNULL = TRUE, {\n    if (is.null(steps.enabled()))\n        rv$steps.status <- setNames(rep(global$UNDONE, rv$length),\n                                 nm = names(rv$config@steps))\n    else\n        rv$steps.status <- steps.status()\n})\n\nobserveEvent(remoteReset(), {\n      lapply(names(rv.widgets), function(x){\n          rv.widgets[[x]] <- widgets.default.values[[x]]\n        })\n})\n\n\n"

Module_Return_Func()
#> [1] "# Return value of module\n# DO NOT MODIFY THIS PART\nlist(config = reactive({\n    config@ll.UI <- setNames(lapply(names(config@steps),\n                                  function(x){\n                                    do.call(\"uiOutput\", list(ns(x)))\n                                  }),\n                           paste0(\"screen_\", names(config@steps))\n    )\n    config\n    }),\ndataOut = reactive({dataOut})\n)\n\n\n"

widgets <- c('widget1', 'widget2')
custom <- list(foo1 = list(), foo2 = 3)
Get_Worflow_Core_Code(widgets, names(custom))
#> [1] "rv.widgets <- reactiveValues(\n\twidget1 = widgets.default.values$widget1,\n\twidget2 = widgets.default.values$widget2\n)\n\n\n    config@steps <- setNames(config@steps,\n        nm = gsub(' ', '', config@steps, fixed = TRUE))\n\n\n    observeEvent(input$widget1, {\n    rv.widgets$widget1 <- input$widget1})\nobserveEvent(input$widget2, {\n    rv.widgets$widget2 <- input$widget2})\n\n\nrv <- reactiveValues(\n    # Stores the object given in input of the process\n    dataIn = NULL,\n    # A vector of boolean indicating the status (UNDONE, SKIPPED or VALIDATED)\n    # of the steps\n    steps.status = NULL,\n    # xxx\n    reset = NULL,\n    # A vector of boolean indicating if the steps are enabled or disabled\n    steps.enabled = NULL\n  )\n\n    rv.custom <- reactiveValues(\n\tfoo1 = rv.custom.default.values$foo1,\n\tfoo2 = rv.custom.default.values$foo2\n)\n\ndataOut <- reactiveValues(\n                trigger = NULL,\n                value = NULL\n                )\n\nobserveEvent(steps.enabled(), ignoreNULL = TRUE, {\n    if (is.null(steps.enabled()))\n        rv$steps.enabled <- setNames(rep(FALSE, rv$length),\n                                 nm = names(rv$config@steps))\n    else\n        rv$steps.enabled <- steps.enabled()\n})\n\nobserveEvent(steps.status(), ignoreNULL = TRUE, {\n    if (is.null(steps.enabled()))\n        rv$steps.status <- setNames(rep(global$UNDONE, rv$length),\n                                 nm = names(rv$config@steps))\n    else\n        rv$steps.status <- steps.status()\n})\n\nobserveEvent(remoteReset(), {\n      lapply(names(rv.widgets), function(x){\n          rv.widgets[[x]] <- widgets.default.values[[x]]\n        })\n})\n\n\n\n"

widgets <- c('widget1', 'widget2')
Get_AdditionalModule_Core_Code(widgets)
#> [1] "rv.widgets <- reactiveValues(\n\twidget1 = widgets.default.values$widget1,\n\twidget2 = widgets.default.values$widget2\n)\n\nobserveEvent(input$widget1, {\n    rv.widgets$widget1 <- input$widget1})\nobserveEvent(input$widget2, {\n    rv.widgets$widget2 <- input$widget2})\n\n\nrv <- reactiveValues(\n    # Stores the object given in input of the process\n    dataIn = NULL,\n    # A vector of boolean indicating the status (UNDONE, SKIPPED or VALIDATED)\n    # of the steps\n    steps.status = NULL,\n    # xxx\n    reset = NULL,\n    # A vector of boolean indicating if the steps are enabled or disabled\n    steps.enabled = NULL\n  )\n\n    rv.custom <- reactiveValues()\n\ndataOut <- reactiveValues(\n                trigger = NULL,\n                value = NULL\n                )\n\n\n\nobserveEvent(reset(), {\n    lapply(names(rv.widgets), function(x){\n        rv.widgets[[x]] <- widgets.default.values[[x]]\n    })\n})\n\n\n\n"

Get_Code_for_AddMod_observeEvents()
#> [1] "\n\nobserveEvent(reset(), {\n    lapply(names(rv.widgets), function(x){\n        rv.widgets[[x]] <- widgets.default.values[[x]]\n    })\n})\n\n\n"